add test and release notes
