+++ pilot/pkg/networking/core/v1alpha3/listener_inbound.go
@@ -659,8 +659,8 @@ func buildSidecarInboundHTTPOpts(lb *ListenerBuilder, cc inboundChainConfig) *ht
 		}
 	}
 
-	if lb.push.Networks != nil {
-		if internalnetwork, exists := lb.push.Networks.Networks[features.InternalAddressMeshNetwork]; exists {
+	if features.EnableHCMInternalNetworks && lb.push.Networks != nil {
+		for _, internalnetwork := range lb.push.Networks.Networks {
 			iac := &hcm.HttpConnectionManager_InternalAddressConfig{}
 			for _, ne := range internalnetwork.Endpoints {
 				if cidr := util.ConvertAddressToCidr(ne.GetFromCidr()); cidr != nil {
